Python/python-microservice/pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = ""
authors = ["Fabricio <fabriciolmd2308@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.6.4"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
requests = "^2.31.0"
aiohttp = "^3.9.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
